feat: added product handlers

+++ internal/dto/product.go
@@ -5,7 +5,7 @@ import "mime/multipart"
 type ProductAtListResponse struct {
 	ID          int            `json:"id"`
 	Name        string         `json:"name"`
-	Weight      int            `json:"weight"`
+	Weight      float64        `json:"weight"`
 	Description string         `json:"desc"`
 	Params      map[string]any `json:"params,omitempty"`
 	Barcode     string         `json:"barcode_url"` // Ссылка на доступ к штрихкоду.
@@ -13,7 +13,7 @@ type ProductAtListResponse struct {
 
 type ProductRequest struct {
 	Name        string         `json:"name"`
-	Weight      *int           `json:"weight"`
+	Weight      *float64       `json:"weight"`
 	Description string         `json:"desc"`
 	Params      map[string]any `json:"params"`
 	Barcode     *Photo         `json:"barcode"` // Штрихкод в байтах
